Split triangle and jigsaw to netcdf into separate files
Add docstrings to each.
