+++ conda_package/mpas_tools/mesh/creation/__init__.py
@@ -1,5 +1,5 @@
-from mpas_tools.mesh.creation.triangle_jigsaw_to_netcdf import \
-    triangle_to_netcdf, jigsaw_to_netcdf
+from mpas_tools.mesh.creation.triangle_to_netcdf import triangle_to_netcdf
+from mpas_tools.mesh.creation.jigsaw_to_netcdf import jigsaw_to_netcdf
 from mpas_tools.mesh.creation.inject_meshDensity import inject_meshDensity
 from mpas_tools.mesh.creation.mpas_to_triangle import mpas_to_triangle
 from mpas_tools.mesh.creation.open_msh import readmsh
