+++ conda_package/recipe/meta.yaml
@@ -19,8 +19,8 @@ build:
     - inject_bathymetry = mpas_tools.mesh.creation.inject_bathymetry:main
     - inject_preserve_floodplain = mpas_tools.mesh.creation.inject_preserve_floodplain:main
     - mpas_to_triangle = mpas_tools.mesh.creation.mpas_to_triangle:main
-    - triangle_to_netcdf = mpas_tools.mesh.creation.triangle_jigsaw_to_netcdf:main_triangle
-    - jigsaw_to_netcdf = mpas_tools.mesh.creation.triangle_jigsaw_to_netcdf:main_jigsaw
+    - triangle_to_netcdf = mpas_tools.mesh.creation.triangle_to_netcdf:main
+    - jigsaw_to_netcdf = mpas_tools.mesh.creation.jigsaw_to_netcdf:main
 
 requirements:
   build:
